+++ CHANGELOG.md
@@ -1,12 +1,13 @@
 # Changelog
 
 ## [Unreleased]
-- Non stated character names as text.
-- Vscode suggestion for normal txt for the text.
-- Documentation shown for dialogic commands.
+- Non stated character names as dialogue text.
+- Make character suggestion part of the normal suggestion in the text.
+- Vscode suggestion similar to txt for dialogue text.
 - Markdown balises highlight.
 - ctrl+click on the jump label moves toward the position of the label.
 - Bug: the color of the choice continues even after the | where it should stop.
+- if `[wait]` is inside dialogue text it will not have good color.
 
 ## [0.1.0] - 2026-09-03
 - Syntax highlight for the important commands.
@@ -14,3 +15,7 @@
 - Suggestion of character names.
 - Most of the balises missed are now highlighted.
 - Bug correction.
+## [0.1.2] - 2026-09-03
+- Add while to the commands.
+- Documentation shown for dialogic commands.
+- Recognize text_input command.